requirements.txt
flask
python-dotenv
redis
dialogflow
python-dotenv